+++ src/main/java/org/example/dto/ProductDTO.java
@@ -1,12 +1,14 @@
 package org.example.dto;
 
+import java.math.BigDecimal;
+
 public class ProductDTO {
     private Long id;
     private String name;
-    private double price;
+    private BigDecimal price;
     private String categoryName;
 
-    public ProductDTO(Long id, String name, double price, String categoryName) {
+    public ProductDTO(Long id, String name, BigDecimal price, String categoryName) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -15,6 +17,6 @@ public class ProductDTO {
 
     public Long getId() { return id; }
     public String getName() { return name; }
-    public double getPrice() { return price; }
+    public BigDecimal getPrice() { return price; }
     public String getCategoryName() { return categoryName; }
 }
